+++ yark/utils.py
@@ -1,8 +1,5 @@
-"""Useful shared utility functions"""
+"""Compatibility exports for shared core utilities."""
 
+from .core.utils import _truncate_text
 
-def _truncate_text(text: str, to: int = 31) -> str:
-    """Truncates inputted `text` to ~32 length, adding ellipsis at the end if overflowing"""
-    if len(text) > to:
-        text = text[: to - 2].strip() + ".."
-    return text.ljust(to)
+__all__ = ["_truncate_text"]
